youtube_downloader/pyproject.toml
[tool.poetry]
name = "youtube-downloader"
version = "0.1.0"
description = ""
authors = ["Dhiraj Suvarna"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
pytube = "^15.0.0"
typer = "^0.9.0"
pydantic = "^2.5.3"
youtube-transcript-api = "^0.6.2"
python-slugify = "^8.0.1"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"